# Add affine channel op to speed and save memory for faster-rcnn model.