… paddle/fluid/operators/CMakeLists.txt
… python/paddle/fluid/tests/unittests/test_affine_channel_op.py