Applies a separate affine transformation to each channel of the input. Useful
for replacing spatial batch norm with its equivalent fixed transformation.

Speed and memory usage for faster-rcnn model:

- speed: 


10个iter时间 | 版本1 | 版本2：generate_proposals加速 | 版本3：affine_channel_op加速
-- | -- | -- | --
1 GPU | 10.843s | 8.663s | 6.701s
8 GPU| 13.663s| 11.315s | 10.532s

- max batch size:

max batch size | 版本1 | 版本2：generate_proposals加速 | 版本3：affine_channel_op加速
-- | -- | -- | --
1 GPU | 1 | 1 | 2
8 GPU| 8| 8| 16